webgl context
now sketch-wrapper supports 2d or webgl context. it was a bit of
headache to do proper type checks. i'm not sure it's complete in terms
of type checking, but it should be okay when using the module (i hope).

in the future,
it might be best to separate context(2d) and gl(webgl). right now,
context could be 2d or webgl, gl could be undefined.
when users use these objects, it should be 100% certain they are
available as needed.