wip - low playFps
the way low playFps works is by early returning loop() without running
render(). problem is when sketch is resized as it bypasses the check and
calls render() directly so the canvas will become visible again.
this will use in-between playhead value and animation shifts a little.

this commit uses snapBy() in advanceTime() to regulate time, but has its
own issues. playhead floors original value, so it's not accurate
representation of time. also, when duration and playFps is both set,
time still flows at normal speed.

still working on it.